Working product

+++ MyServer.Application/DTOs/Patients/SexDto.cs
@@ -2,7 +2,7 @@
 {
     public class SexDto
     {
-        public int Code { get; set; }
+        public short Code { get; set; }
         public string Identifier { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
     }
